% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curves.R
\name{compute_window_slope}
\alias{compute_window_slope}
\title{Compute the slope of the confluency in a window}
\usage{
compute_window_slope(experiment, winsize = 20, do_plot = TRUE)
}
\arguments{
\item{experiment}{An experiment tibble as returned by process_growth_curve}

\item{winsize}{The half-size of the window in time units}

\item{do_plot}{Boolean whether the slopes microplate should be plotted}
}
\value{
Invisibly the experiment tibble without the first 'winsize' data points and an extra column 'Slope'.
}
\description{
Compute the slope of the confluency in a window using the max-min approximation. Does the direct sum, considering the values are log-transformed
}
