% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{concentration_values}
\alias{concentration_values}
\title{Convert readable concentration format to numerals}
\usage{
concentration_values(concentrations)
}
\arguments{
\item{concentrations}{A vector of readable concentrations}
}
\value{
A numeric vector corresponding to the converted values
}
\description{
Convert from the readable "XX[pnµm]M" format to a numeral M format
}
