\name{drugResistance-package}
\alias{drugResistance-package}
\alias{drugResistance}
\docType{package}
\title{
\packageTitle{drugResistance}
}
\description{
\packageDescription{drugResistance}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{drugResistance}
\packageIndices{drugResistance}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{drugResistance}

Maintainer: \packageMaintainer{drugResistance}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
