% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_effective_concentration}
\alias{find_effective_concentration}
\alias{findIC20}
\title{Find the concentration for a given inhibition}
\usage{
find_effective_concentration(
  nls_model,
  target_effect = 0.5,
  explore = c(-9:1),
  precision = 0.01
)

findIC20(nls_model, explore = c(-9:1), precision = 0.01)
}
\arguments{
\item{nls_model}{Model fitted by fit_drug_sensitivity}

\item{target_effect}{Percent viability that should be reached}

\item{explore}{Starting range to look for the concentration, in log10}

\item{precision}{How close to the target effect can the algorithm stop}
}
\description{
Find the target concentration by predicting the effect on a range of concentration and shrinking from there
}
