% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curves.R
\name{fit_drug_sensitivity}
\alias{fit_drug_sensitivity}
\title{Fit sigmoid to drug sensitivities}
\usage{
fit_drug_sensitivity(pexp, controls = c("DMSO", "control", "medium",
  "cell"))
}
\description{
Fit sigmoid to drug sensitivities
}
