% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fit_sigmoid}
\alias{fit_sigmoid}
\title{Fit a sigmoid}
\usage{
fit_sigmoid(fit_data)
}
\arguments{
\item{fit_data}{A tibble or data.frame with columns 'Concentration_value' and 'Viability' corresponding defining the points to fit}
}
\value{
A vector of size 3 corresponding to parameters b and c, and the residual of the fit
}
\description{
Helper function to fit a sigmoid with a=1 to data using the LM algorithm
}
\seealso{
sigmoid
}
