% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curves.R
\name{microplate}
\alias{microplate}
\title{Convert a well ID into X and Y coordinates}
\usage{
microplate(input_tibble)
}
\arguments{
\item{input_tibble}{A tibble with column 'Well' matching the regex '[A-Z]\\d\\+' (one letter and a number)
export}
}
\description{
Convert a well ID into X and Y coordinates that are factors and probably ordered for facet_wrap
}
