% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_calculations.R
\name{plot_bliss_scores}
\alias{plot_bliss_scores}
\title{Plot starHeatmaps for a group of synergy data}
\usage{
plot_bliss_scores(pdata, restrict = TRUE, col_drug = "", control = "DMSO")
}
\arguments{
\item{pdata}{Processed data}
}
\description{
Plot starHeatmaps for a group of synergy data
}
