% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curves.R
\name{plot_growth_rates}
\alias{plot_growth_rates}
\title{Plot processed growth rates}
\usage{
plot_growth_rates(pgc)
}
\arguments{
\item{pgc}{A tibble as outputed by process_growth_curve}
}
\description{
Plot a heatmap of the mean growth rate in each well of a plate.
}
