% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curves.R
\name{plot_ic50_fit}
\alias{plot_ic50_fit}
\title{Plot the drug sensitivity with the fit}
\usage{
plot_ic50_fit(fit_result, drug_name = "")
}
\arguments{
\item{fit_result}{Result of fit_drug_sensitivity}

\item{drug_name}{Drug for which the plot should be made}
}
\description{
Plot the drug sensitivity with the fit
}
