% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curves.R
\name{process_single_growth_curve}
\alias{process_single_growth_curve}
\title{Process one growth curve to a viability value}
\usage{
process_single_growth_curve(trace)
}
\arguments{
\item{trace}{A tibble with columns Elapsed for the time and Value for the confluency at those times}
}
\value{
A one line tibble with the same columns where value is now the growth rate
}
\description{
Process one growth curve to a viability value
}
