% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sigmoid}
\alias{sigmoid}
\title{The sigmoid function}
\usage{
sigmoid(params, xx)
}
\arguments{
\item{params}{A vector of size 2 for the 'b' and 'c' parameters}

\item{xx}{A vector of numerals to which the function should be applied}
}
\value{
A vector of numerals
}
\description{
The sigmoid function: a / (1 + exp(-b*(x-c)))
}
