% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_calculations.R
\name{starHeatmap2}
\alias{starHeatmap2}
\alias{starHeatmap}
\title{Heatmap with stars for extra data}
\usage{
starHeatmap2(hm_values, data, encoding = "size", return_list = FALSE, ...)

starHeatmap(data, ...)
}
\arguments{
\item{hm_values}{Values for the colors of the heatmap tiles}

\item{data}{Values for the text (or number of stars) of the heatmap tiles}

\item{encoding}{One of c('size', 'count', 'multi_count') for whether the value in data should be encoded by the size of one star or a number of stars on one or multiple lines}

\item{return_list}{If FALSE the heatmap is plotted with a legend, if TRUE returns a list with fields 'object' for the Heatmap object and 'annotation_legend_list' for the annotation corresponding to the encoding.}
}
\description{
Plot heatmap with stars in the tile to encode for another data (significance, effect size, etc)

Helper for starHeatmap2 with hm_values = data
}
