% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_calculations.R
\name{starHeatmap2}
\alias{starHeatmap2}
\title{Heatmap with stars for extra data}
\usage{
starHeatmap2(hm_values, data, encoding = "size", ...)
}
\arguments{
\item{hm_values}{Values for the colors of the heatmap tiles}

\item{data}{Values for the text (or number of stars) of the heatmap tiles}

\item{encoding}{One of c('size', 'count', 'multi_count') for whether the value in data should be encoded by the size of one star or a number of stars on one or multiple lines}
}
\description{
Plot heatmap with stars in the tile to encode for another data (significance, effect size, etc)
}
