% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy_calculations.R
\name{bliss_score}
\alias{bliss_score}
\title{Helper to get bliss score matrix}
\usage{
bliss_score(pdata, restrict = TRUE)
}
\description{
Helper to get a nice bliss score matrix from processed data
}
