% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_curves.R
\name{process_growth_curve}
\alias{process_growth_curve}
\title{Process growth curves to viability values}
\usage{
process_growth_curve(
  experiment,
  keep_controls = TRUE,
  max_confluency = 75,
  space = log
)
}
\arguments{
\item{experiment}{Dataset as return by read_incu from incucyter}

\item{keep_controls}{Whether the controls should be present in the output. Control values are DMSO, medium and no_cells.}

\item{max_confluency}{The maximum value to consider valid. Used to remove data points when the growth tappers out because of contact inhibition. Adjust according to your data type.}

\item{space}{A function to modify the fitting space (log (default), identity, etc)}
}
\value{
A tibble with new columns Inhibitor, Concentration and Viability
}
\description{
Extract growth rate, normalise and averages growth curves from the experiment
}
